Move away from Page.js to weasyprint
Weasyprint has more feature pairity with the print CSS proposal. It
supports setting a CSS string from a data attribue. Which is the main
change we made here. We also allow for custom setting of the font size,
because for some reason weasyprints font size for "pt" fonts is
different from Page.js. I'm not sure which one is right, but we can
still adjust the size as needed to get it to "one page per page."