Refactor page combiner to take CSS
It became apparent that parametrizing every value in the CSS is
intractable. Instead of making things like page width and fonts
parameters, users will pass in a whole CSS file. We will have to
document that classes exposed, but for now, people can use a manifest in
the manifests folder as a reference.